+++ nc2zarr/preprocessor.py
@@ -76,10 +76,13 @@ def ensure_dataset_has_concat_dim(ds: xr.Dataset,
 
     if concat_dim_var is not None:
         if not concat_dim_var.dims:
-            # if the concat_dim_var does not yet have a dimension, add it
-            ds = ds.assign_coords({
-                concat_dim_name: xr.DataArray(concat_dim_var, dims=(concat_dim_name,))
-            })
+            # If the concat_dim_var does not yet has a dimension, add it.
+            # This is typically the case if time value is a scalar rather
+            # than an 1-element array.
+            concat_dim_var = xr.DataArray([concat_dim_var.values],
+                                          dims=(concat_dim_name,),
+                                          attrs=concat_dim_var.attrs)
+            ds = ds.assign_coords({concat_dim_name: concat_dim_var})
     elif concat_dim_name == 'time':
         time_coverage_start, time_coverage_end = \
             get_time_coverage_from_ds(ds, datetime_format=datetime_format)